fix(data): add deleted rows back to `patient` to prevent misalignment of patient IDs (#584)
In #582, rows were removed from the QuestionnaireDB `patient` table, to let the `SyncPublishQuestionnaire` event populate this table.

However, this caused a misalignment of patient IDs, because deleted patients were no longer being added to the table. This led to rows in `answerQuestionnaire` and `answer` being misattributed to the wrong users (e.g., a row in `answerQuestionnaire` previously attributing a questionnaire to Laurie, patient 9, was now linking it to Kathy, new patient 9).

This PR adds all row inserts back to `patient` (including the deleted rows), to fix the offset in IDs (e.g. restoring Laurie as patient 9), and to prevent issues with unpredictable syncing in general. This will also allow us to more easily clean up questionnaire data for deleted users.