fix(data): add back those test patient and caregivers that are in test data (#578)
Inserts back the patients and caregivers that we have in our test data due to problems when this is done in the insert_test_data command (due to truncating).

Also, this allows a user to set up a local environment without requiring to run the insert_test_data management command in admin first.

This should be easier to manage for users in terms of the process to set up that is in various places (README here, in admin, and the overall instructions in our documentation).